+++ R/gpfield-package.R
@@ -1,4 +1,5 @@
 #' @keywords internal
+#' @importFrom stats predict
 "_PACKAGE"
 
 ## usethis namespace: start